Create a sidebar service class
